ulang

+++ routes/web.php
@@ -30,16 +30,16 @@ Route::group(["middleware"=>["dashboard.access"], "namespace"=>"App\Http\Control
 	});
 
     Route::group(["as"=>"menu.", "prefix"=>"menu"], function(){
-        Route::get("/", "MenuController@index")->name("index");
+        Route::get("/", "MenuController@index")->name("index")->middleware('role:'. implode('|', [RoleEnum::SuperAdmin, RoleEnum::Admin]));
     });
 
     Route::group(["as"=>"informasi.", "prefix"=>"informasi"], function(){
-        Route::get("/", "InformasiController@index")->name("index");
-        Route::get("/create", "InformasiController@create")->name("create");
-        Route::post("/","InformasiController@store")->name("store");
-        Route::get('/{id}/edit', 'InformasiController@edit')->name("edit");
-        Route::put('/{id}', 'InformasiController@update')->name("update");
-        Route::delete('/{id}', 'InformasiController@destroy')->name("destroy");
+        Route::get("/", "InformasiController@index")->name("index")->middleware('role:'. implode('|', [RoleEnum::SuperAdmin, RoleEnum::Admin]));
+        Route::get("/create", "InformasiController@create")->name("create")->middleware('role:'. implode('|', [RoleEnum::SuperAdmin, RoleEnum::Admin]));
+        Route::post("/","InformasiController@store")->name("store")->middleware('role:'. implode('|', [RoleEnum::SuperAdmin, RoleEnum::Admin]));
+        Route::get('/{id}/edit', 'InformasiController@edit')->name("edit")->middleware('role:'. implode('|', [RoleEnum::SuperAdmin, RoleEnum::Admin]));
+        Route::put('/{id}', 'InformasiController@update')->name("update")->middleware('role:'. implode('|', [RoleEnum::SuperAdmin, RoleEnum::Admin]));
+        Route::delete('/{id}', 'InformasiController@destroy')->name("destroy")->middleware('role:'. implode('|', [RoleEnum::SuperAdmin, RoleEnum::Admin]));
     });
 
     Route::group(["as" => "berita.","prefix" => "berita"], function () {
@@ -53,86 +53,78 @@ Route::group(["middleware"=>["dashboard.access"], "namespace"=>"App\Http\Control
 	});
 
     Route::group(["as" => "banner.","prefix" => "banner"], function () {
-		Route::get('/', 'BannerController@index')->name("index");
-		Route::get('/create', 'BannerController@create')->name("create");
-        Route::post('/', 'BannerController@store')->name("store");
-		Route::get('/{id}', 'BannerController@show')->name("show");
-		Route::get('/{id}/edit', 'BannerController@edit')->name("edit");
-		Route::put('/{id}', 'BannerController@update')->name("update");
-		Route::delete('/{id}', 'BannerController@destroy')->name("destroy");
+		Route::get('/', 'BannerController@index')->name("index")->middleware('role:'. implode('|', [RoleEnum::SuperAdmin, RoleEnum::Admin]));
+		Route::get('/create', 'BannerController@create')->name("create")->middleware('role:'. implode('|', [RoleEnum::SuperAdmin, RoleEnum::Admin]));
+        Route::post('/', 'BannerController@store')->name("store")->middleware('role:'. implode('|', [RoleEnum::SuperAdmin, RoleEnum::Admin]));
+		Route::get('/{id}', 'BannerController@show')->name("show")->middleware('role:'. implode('|', [RoleEnum::SuperAdmin, RoleEnum::Admin]));
+		Route::get('/{id}/edit', 'BannerController@edit')->name("edit")->middleware('role:'. implode('|', [RoleEnum::SuperAdmin, RoleEnum::Admin]));
+		Route::put('/{id}', 'BannerController@update')->name("update")->middleware('role:'. implode('|', [RoleEnum::SuperAdmin, RoleEnum::Admin]));
+		Route::delete('/{id}', 'BannerController@destroy')->name("destroy")->middleware('role:'. implode('|', [RoleEnum::SuperAdmin, RoleEnum::Admin]));
 	});
 
-    Route::group(["as"=>"page.", "prefix"=>"page"], function(){
-        Route::get("/", "PageController@index")->name("index");
-    });
-
-    Route::group(["as"=>"berita.", "prefix"=>"berita"], function(){
-        Route::get("/", "BeritaController@index")->name("index");
-    });
-
     Route::group(["as"=>"galeri.", "prefix"=>"galeri"], function(){
-        Route::get("/", "GaleriController@index")->name("index");
-        Route::get("/create", "GaleriController@create")->name("create");
-        Route::post("/","GaleriController@store")->name("store");
-        Route::get('/{id}', 'GaleriController@show')->name("show");
-        Route::get('/{id}/edit', 'GaleriController@edit')->name("edit");
-        Route::put('/{id}', 'GaleriController@update')->name("update");
-        Route::delete('/{id}', 'GaleriController@destroy')->name("destroy");
+        Route::get("/", "GaleriController@index")->name("index")->middleware('role:'. implode('|', [RoleEnum::SuperAdmin, RoleEnum::Admin]));
+        Route::get("/create", "GaleriController@create")->name("create")->middleware('role:'. implode('|', [RoleEnum::SuperAdmin, RoleEnum::Admin]));
+        Route::post("/","GaleriController@store")->name("store")->middleware('role:'. implode('|', [RoleEnum::SuperAdmin, RoleEnum::Admin]));
+        Route::get('/{id}', 'GaleriController@show')->name("show")->middleware('role:'. implode('|', [RoleEnum::SuperAdmin, RoleEnum::Admin]));
+        Route::get('/{id}/edit', 'GaleriController@edit')->name("edit")->middleware('role:'. implode('|', [RoleEnum::SuperAdmin, RoleEnum::Admin]));
+        Route::put('/{id}', 'GaleriController@update')->name("update")->middleware('role:'. implode('|', [RoleEnum::SuperAdmin, RoleEnum::Admin]));
+        Route::delete('/{id}', 'GaleriController@destroy')->name("destroy")->middleware('role:'. implode('|', [RoleEnum::SuperAdmin, RoleEnum::Admin]));
     });
 
     Route::group(["as"=>"kalender.", "prefix"=>"kalender"], function(){
-        Route::get("/", "KalenderController@index")->name("index");
-        Route::get("/create", "KalenderController@create")->name("create");
-        Route::post("/","KalenderController@store")->name("store");
-        Route::get('/{id}', 'KalenderController@show')->name("show");
-        Route::get('/{id}/edit', 'KalenderController@edit')->name("edit");
-        Route::put('/{id}', 'KalenderController@update')->name("update");
-        Route::delete('/{id}', 'KalenderController@destroy')->name("destroy");
+        Route::get("/", "KalenderController@index")->name("index")->middleware('role:'. implode('|', [RoleEnum::SuperAdmin, RoleEnum::Admin]));
+        Route::get("/create", "KalenderController@create")->name("create")->middleware('role:'. implode('|', [RoleEnum::SuperAdmin, RoleEnum::Admin]));
+        Route::post("/","KalenderController@store")->name("store")->middleware('role:'. implode('|', [RoleEnum::SuperAdmin, RoleEnum::Admin]));
+        Route::get('/{id}', 'KalenderController@show')->name("show")->middleware('role:'. implode('|', [RoleEnum::SuperAdmin, RoleEnum::Admin]));
+        Route::get('/{id}/edit', 'KalenderController@edit')->name("edit")->middleware('role:'. implode('|', [RoleEnum::SuperAdmin, RoleEnum::Admin]));
+        Route::put('/{id}', 'KalenderController@update')->name("update")->middleware('role:'. implode('|', [RoleEnum::SuperAdmin, RoleEnum::Admin]));
+        Route::delete('/{id}', 'KalenderController@destroy')->name("destroy")->middleware('role:'. implode('|', [RoleEnum::SuperAdmin, RoleEnum::Admin]));
     });
 
     Route::group(["as"=>"kontak.", "prefix"=>"kontak"], function(){
-        Route::get("/", "KontakController@index")->name("index");
+        Route::get("/", "KontakController@index")->name("index")->middleware('role:'. implode('|', [RoleEnum::SuperAdmin, RoleEnum::Admin]));
     });
 
     Route::group(["as" => "layanan.","prefix" => "layanan"], function () {
-		Route::get('/', 'LayananController@index')->name("index");
-		Route::get('/create', 'LayananController@create')->name("create");
-        Route::post('/', 'LayananController@store')->name("store");
-		Route::get('/{id}', 'LayananController@show')->name("show");
-		Route::get('/{id}/edit', 'LayananController@edit')->name("edit");
-		Route::put('/{id}', 'LayananController@update')->name("update");
-		Route::delete('/{id}', 'LayananController@destroy')->name("destroy");
+		Route::get('/', 'LayananController@index')->name("index")->middleware('role:'. implode('|', [RoleEnum::SuperAdmin, RoleEnum::Admin]));
+		Route::get('/create', 'LayananController@create')->name("create")->middleware('role:'. implode('|', [RoleEnum::SuperAdmin, RoleEnum::Admin]));
+        Route::post('/', 'LayananController@store')->name("store")->middleware('role:'. implode('|', [RoleEnum::SuperAdmin, RoleEnum::Admin]));
+		Route::get('/{id}', 'LayananController@show')->name("show")->middleware('role:'. implode('|', [RoleEnum::SuperAdmin, RoleEnum::Admin]));
+		Route::get('/{id}/edit', 'LayananController@edit')->name("edit")->middleware('role:'. implode('|', [RoleEnum::SuperAdmin, RoleEnum::Admin]));
+		Route::put('/{id}', 'LayananController@update')->name("update")->middleware('role:'. implode('|', [RoleEnum::SuperAdmin, RoleEnum::Admin]));
+		Route::delete('/{id}', 'LayananController@destroy')->name("destroy")->middleware('role:'. implode('|', [RoleEnum::SuperAdmin, RoleEnum::Admin]));
 	});
 
     Route::group(["as"=>"tautan.", "prefix"=>"tautan"], function(){
-        Route::get("/", "TautanController@index")->name("index");
-        Route::get("/create", "TautanController@create")->name("create");
-        Route::post("/","TautanController@store")->name("store");
-        Route::get('/{id}', 'TautanController@show')->name("show");
-        Route::get('/{id}/edit', 'TautanController@edit')->name("edit");
-        Route::put('/{id}', 'TautanController@update')->name("update");
-        Route::delete('/{id}', 'TautanController@destroy')->name("destroy");
+        Route::get("/", "TautanController@index")->name("index")->middleware('role:'. implode('|', [RoleEnum::SuperAdmin, RoleEnum::Admin]));
+        Route::get("/create", "TautanController@create")->name("create")->middleware('role:'. implode('|', [RoleEnum::SuperAdmin, RoleEnum::Admin]));
+        Route::post("/","TautanController@store")->name("store")->middleware('role:'. implode('|', [RoleEnum::SuperAdmin, RoleEnum::Admin]));
+        Route::get('/{id}', 'TautanController@show')->name("show")->middleware('role:'. implode('|', [RoleEnum::SuperAdmin, RoleEnum::Admin]));
+        Route::get('/{id}/edit', 'TautanController@edit')->name("edit")->middleware('role:'. implode('|', [RoleEnum::SuperAdmin, RoleEnum::Admin]));
+        Route::put('/{id}', 'TautanController@update')->name("update")->middleware('role:'. implode('|', [RoleEnum::SuperAdmin, RoleEnum::Admin]));
+        Route::delete('/{id}', 'TautanController@destroy')->name("destroy")->middleware('role:'. implode('|', [RoleEnum::SuperAdmin, RoleEnum::Admin]));
     });
 
 
     Route::group(["as"=>"pengaturan.", "prefix"=>"pengaturan"], function(){
-        Route::get("/", "PengaturanController@index")->name("index");
+        Route::get("/", "PengaturanController@index")->name("index")->middleware('role:'. implode('|', [RoleEnum::SuperAdmin]));
     });
 
     Route::group(["as"=>"log.", "prefix"=>"log"], function(){
-        Route::get("/", "LogController@index")->name("index");
+        Route::get("/", "LogController@index")->name("index")->middleware('role:'. implode('|', [RoleEnum::SuperAdmin]));
     });
 
     Route::group(["as"=>"users.", "prefix"=>"users"], function(){
-        Route::get("/", "UserController@index")->name("index")->middleware('role:SuperAdmin');
-        Route::get('/create', 'UserController@create')->name("create")->middleware('role:SuperAdmin');
-        Route::post('/', 'UserController@store')->name("store")->middleware('role:SuperAdmin');
-        Route::get('/{id}', 'UserController@show')->name("show")->middleware('role:SuperAdmin');
-        Route::get('/{id}/edit', 'UserController@edit')->name("edit")->middleware('role:SuperAdmin');
-        Route::put('/{id}', 'UserController@update')->name("update")->middleware('role:SuperAdmin');
-        Route::delete('/{id}', 'UserController@destroy')->name("destroy")->middleware('role:SuperAdmin');
-        Route::patch('/{id}', 'UserController@restore')->name("restore")->middleware('role:SuperAdmin');
-        Route::get('/{id}/impersonate', 'UserController@impersonate')->name("impersonate")->middleware('role:SuperAdmin');
+        Route::get("/", "UserController@index")->name("index")->middleware('role:'. implode('|', [RoleEnum::SuperAdmin]));
+        Route::get('/create', 'UserController@create')->name("create")->middleware('role:'. implode('|', [RoleEnum::SuperAdmin]));
+        Route::post('/', 'UserController@store')->name("store")->middleware('role:'. implode('|', [RoleEnum::SuperAdmin]));
+        Route::get('/{id}', 'UserController@show')->name("show")->middleware('role:'. implode('|', [RoleEnum::SuperAdmin]));
+        Route::get('/{id}/edit', 'UserController@edit')->name("edit")->middleware('role:'. implode('|', [RoleEnum::SuperAdmin]));
+        Route::put('/{id}', 'UserController@update')->name("update")->middleware('role:'. implode('|', [RoleEnum::SuperAdmin]));
+        Route::delete('/{id}', 'UserController@destroy')->name("destroy")->middleware('role:'. implode('|', [RoleEnum::SuperAdmin]));
+        Route::patch('/{id}', 'UserController@restore')->name("restore")->middleware('role:'. implode('|', [RoleEnum::SuperAdmin]));
+        Route::get('/{id}/impersonate', 'UserController@impersonate')->name("impersonate")->middleware('role:'. implode('|', [RoleEnum::SuperAdmin]));
     });
     
 });
@@ -168,6 +160,7 @@ Route::group(["namespace"=>"App\Http\Controllers\Home", "as"=>"home."], function
 
     Route::group(["as"=>"berita.", "prefix"=>"berita"], function(){
         Route::get("/", "BeritaController@index")->name("index");
+        Route::get("/{id}", "BeritaController@show")->name("show");
     });
 
     Route::group(["as"=>"inovasi.", "prefix"=>"inovasi"], function(){
